Make code generation work correctly in Julia v1.9
The main problem was that the `ccall` to the extern `getVec` function was trying
to return either an `i64` or `i32`, depending on the value of one of the
arguments (`i64` for the pointer of the array, `i32` for its size).  The
solution was to split this function into two, one returning `uintptr_t`/`i64`,
and the other returning `uint32_t`/`i32`, so that the `ccall` can have a more
straightforward signature.

Another problem that arise in Julia v1.9 was that the new default algorithm of
`sort!` requires a scratch array, which we can't allocate on the IPU, so we need
to use the simpler `QuickSort` algorithm.  Also, to use this function we had to
increase the stack size of the kernel in the codelet.

Fix #8.