Add a mechanism to disable printing in kernels with `@ipuprint*` macros
Luk Burchard suggested that having code for printing to the IPU might be costly,
even if the print is not actually run, but later evaluation showed that the cost
should be negligible.  In any case it is easy enough to add the mechanism to
disable printing, so let's do it.

Fix #11.