Add function to release all attached devices
In the original version of the code developed by Emily and Luk there was a
global list of attached devices, but it wasn't protected with a lock, and there
wasn't a specific function to detach all devices.  This commit brings back the
idea, expanding it and making it safer and more useful.